backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
requires-python = ">=3.13.4,<3.14"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "python-multipart<1.0.0,>=0.0.21",
    "email-validator<3.0.0.0,>=2.3.0",
    "tenacity<10.0.0,>=9.1.2",
    "pydantic<3.0.0,>=2.12.5",
    "emails<1.0,>=0.6",
    "jinja2<4.0.0,>=3.1.6",
    "alembic<2.0.0,>=1.18.1",
    "httpx<1.0.0,>=0.28.1",
    "psycopg[binary]<4.0.0,>=3.3.2",
    "sqlmodel<1.0.0,>=0.0.31",
    "bcrypt==5.0.0",
    "pydantic-settings<3.0.0,>=2.12.0",
    "sentry-sdk[fastapi]<3.0.0,>=2.50.0",
    "pyjwt<3.0.0,>=2.10.1",
    "firebase-admin>=7.1.0,<8.0.0",
    "qstash>=3.2.0",
    "maileroo>=1.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest<10.0.0,>=9.0.2",
    "mypy<2.0.0,>=1.19.1",
    "ruff<1.0.0,>=0.14.14",
    "pre-commit<5.0.0,>=4.5.1",
    "coverage<8.0.0,>=7.13.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic"]

[tool.ruff]
target-version = "py313"
exclude = ["alembic"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG001", # unused arguments in functions
    "T201",   # print statements are not allowed
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true

[tool.coverage.run]
source = ["app"]
dynamic_context = "test_function"

[tool.coverage.report]
show_missing = true
sort = "-Cover"

[tool.coverage.html]
show_contexts = true
